triage-stage1/requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy>=2.0
alembic
psycopg[binary]
pydantic>=2.0
pillow
markdown-it-py
bleach
argon2-cffi
httpx
python-multipart
pytest
